🔨 Update to use `build` for package building
Still using `setuptools` as the build backend, but switch to
https://github.com/pypa/build for the frontend (it builds both sdist and
wheel in the same command, and is PEP 517).

Several related updates:

- the scripts/dtrx file is moved to package layout (dtrx/dtrx.py), but
  is still runnable standalone
- reworking the tree to a python package style enables simpler and more
  typical package setup
- switch to `setup.cfg`, reduces the amount of messiness in the build
  configuration
- move the original documentation to `archived/` to reduce confusion
- add some information about how the package is built + released to
  README.md
- add a Build step to the github actions

The `setup.cfg` file was formatted with the quite nice `setup-cfg-fmt`
package (https://github.com/asottile/setup-cfg-fmt).

This also removes `twine` from the setup-requires dependencies.
Fixes #19.

And moving the package configuration to `setup.cfg`, I noticed the
`subprocess32` backport dependency was not specified correctly when
building the wheel (it was set based on the _build_ environment, not the
install environment), fixed now.