Gentler exit on missing 'file' command
`dtrx` exits quite abruptly with a traceback if the `file` command is
not present on the host (for example, on the docker image
`python:3.10-slim-buster`).

Add a fix to more gently exit with a message on this event:

```bash
dtrx: ERROR: 'file' command not found, skipping magic test
```

Add a pretty janky test to ensure it stays fixed.

Fixes #24