Send stdout to /dev/null unless it is needed
'unrar' produces lots of output to stdout, and previously this
output was stored in a pipe with Subprocess.PIPE, which could get
filled up, causing everything to hang.

Almost all the of the time we do not care what uncompressors are
producing on stdout, so just send it to /dev/null. Add a way of
stopping this behaviour for the one uncompressor which currently
does read the output pipe.